[BugFix] use beijing time instead of local time
The UTC time obtained locally is determined by the system time and
cannot obtain the real Beijing time. Therefore, it is modified to obtain
the standard time from NTP. After that, we will correct the local time
by adding a burden to the current time to ensure the accurate operation
of the script.

issue: #0